requirements.txt
praw==7.7.1
google-generativeai==0.3.2
python-dotenv==1.0.1
requests==2.31.0
tqdm==4.66.1 